# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "No-Code Workflow Studio API",
    "description": "Backend for the **No-Code Workflow Studio** \u2014 build ML workflows on a visual\ncanvas and execute them as a durable DAG against Tapis-managed HPC resources.\n\n### Authentication\n\nEvery endpoint except `GET /` requires an authenticated user. Two credentials\nare accepted, checked in this order:\n\n1. **`X-Tapis-Token`** header \u2014 a Tapis access token. Used when the Studio is\n   embedded in TapisUI, where the host supplies the token.\n2. **`session`** cookie \u2014 set by this app's own Tapis OAuth login\n   (`GET /login` \u2192 `GET /oauth2/callback`).\n\nA token that fails verification falls through to the session rather than\nhard-failing. `GET /auth-debug` reports which credential a request carried and\nwhy it was or wasn't accepted.\n\n### Ownership\n\nTemplates and runs are scoped to the calling user.\n\n* **Runs** are owner-only. Another user's run returns `404`.\n* **Templates** are owner-only until published (`POST\n  /api/workflow-templates/{template_id}/publish`), which grants every user\n  read, run and clone access \u2014 never write. Only the owner can add a version.\n\nResources you may not access return **`404`, not `403`**, so a response never\nconfirms that an id exists.\n",
    "contact": {
      "name": "ICICLE-ai / workflow-orchestrator",
      "url": "https://github.com/ICICLE-ai/workflow-orchestrator"
    },
    "license": {
      "name": "MIT"
...
  "paths": {
...
    "/api/pipeline-runs/{run_id}/geospatial/labels": {
      "get": {
        "tags": [
          "geospatial"
        ],
        "summary": "List Labels",
        "description": "Available layer names in the run's GeoPackage.",
        "operationId": "list_labels_api_pipeline_runs__run_id__geospatial_labels_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated \u2014 no valid X-Tapis-Token or session cookie."
          },
          "404": {
            "description": "Not found, or not visible to the calling user."
          }
        },
        "security": [
          {
            "tapisToken": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/pipeline-runs/{run_id}/geospatial/config": {
      "get": {
        "tags": [
          "geospatial"
        ],
        "summary": "Get Config",
        "description": "Generator/spray-mode config metadata for the step that produced this run's\nGeoPackage (generatorMode, sprayMode, levels, thresholds). A static\n'source_geopackage' node has no such fields \u2014 its config_schema has none of\nthese keys, so every value here comes back null.",
        "operationId": "get_config_api_pipeline_runs__run_id__geospatial_config_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated \u2014 no valid X-Tapis-Token or session cookie."
          },
          "404": {
            "description": "Not found, or not visible to the calling user."
          }
        },
        "security": [
          {
            "tapisToken": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/pipeline-runs/{run_id}/geospatial/missing": {
      "get": {
        "tags": [
          "geospatial"
        ],
        "summary": "List Missing",
        "description": "Expected entries with no corresponding source imagery/data.\n\nASSUMPTION (unconfirmed against the real Tapis app, which lives outside this\nrepo): entries are recorded as a non-spatial attribute layer literally named\n'missing' inside the GeoPackage. If no such layer exists, an empty list is\nreturned rather than treated as an error.",
        "operationId": "list_missing_api_pipeline_runs__run_id__geospatial_missing_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated \u2014 no valid X-Tapis-Token or session cookie."
          },
          "404": {
            "description": "Not found, or not visible to the calling user."
          }
        },
        "security": [
          {
            "tapisToken": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/pipeline-runs/{run_id}/geospatial/geojson/{label}": {
      "get": {
        "tags": [
          "geospatial"
        ],
        "summary": "Get Geojson",
        "description": "Read a GeoPackage layer and convert it to GeoJSON on demand \u2014 this is what\nthe Leaflet map consumes; raw GeoPackage bytes are never returned here.",
        "operationId": "get_geojson_api_pipeline_runs__run_id__geospatial_geojson__label__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "label",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated \u2014 no valid X-Tapis-Token or session cookie."
          },
          "404": {
            "description": "Not found, or not visible to the calling user."
          }
        },
        "security": [
          {
            "tapisToken": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/pipeline-runs/{run_id}/geospatial/download/{label}": {
      "get": {
        "tags": [
          "geospatial"
        ],
        "summary": "Download Shapefile",
        "description": "Convert a GeoPackage layer to a zipped ESRI Shapefile bundle on demand.",
        "operationId": "download_shapefile_api_pipeline_runs__run_id__geospatial_download__label__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "label",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated \u2014 no valid X-Tapis-Token or session cookie."
          },
          "404": {
            "description": "Not found, or not visible to the calling user."
          }
        },
        "security": [
          {
            "tapisToken": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/pipeline-runs/{run_id}/geospatial/download-farm/{label}": {
      "get": {
        "tags": [
          "geospatial"
        ],
        "summary": "Download Farm Boundary",
        "description": "Same conversion as /download, for the farm/field boundary layer.\n\nASSUMPTION (unconfirmed): {label} names the same GeoPackage layer as\n/download/{label} \u2014 this endpoint only changes the output filename\nconvention. If boundary geometry actually lives under one fixed layer name\nindependent of {label}, point this at that fixed name once the real\nGeoPackage schema (produced by the Tapis app outside this repo) is known.",
        "operationId": "download_farm_boundary_api_pipeline_runs__run_id__geospatial_download_farm__label__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "label",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated \u2014 no valid X-Tapis-Token or session cookie."
          },
          "404": {
            "description": "Not found, or not visible to the calling user."
          }
        },
        "security": [
          {
            "tapisToken": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/pipeline-runs/{run_id}/geospatial/download-gpkg/{label}": {
      "get": {
        "tags": [
          "geospatial"
        ],
        "summary": "Download Gpkg",
        "description": "Stream the run's raw GeoPackage file back unmodified, for QGIS/ArcGIS.\n\nA GeoPackage is one file holding every layer, so {label} does not subset\nit \u2014 the whole file is returned regardless of which label was requested;\n{label} only names the downloaded file.",
        "operationId": "download_gpkg_api_pipeline_runs__run_id__geospatial_download_gpkg__label__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "label",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated \u2014 no valid X-Tapis-Token or session cookie."
          },
          "404": {
            "description": "Not found, or not visible to the calling user."
          }
        },
        "security": [
          {
            "tapisToken": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "tapisToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tapis-Token",
        "description": "Tapis access token. Preferred when the Studio is embedded in TapisUI, which supplies it."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Signed session cookie from this app's own Tapis OAuth login (GET /login)."
      }
    }
  },
...
}